pyproject.toml
[project]
name = "urlshutoku"
version = "0.1.0"
description = "Company homepage and metadata collector"
authors = [{ name = "URLshutoku Maintainers", email = "maintainers@example.com" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "beautifulsoup4",
    "python-dotenv",
    "python-multipart",
    "sqlalchemy",
    "psycopg2-binary",
    "jinja2",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rapidfuzz",
    "loguru",
]

[project.optional-dependencies]
llm = [
    "llama-cpp-python>=0.2.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
